requirements.txt
tkinterdnd2
Pillow
customtkinter
torch
pyyaml
opencv-python
numpy